+++ src/thor/astar_bss.cc
@@ -147,7 +147,7 @@ void AStarBSSAlgorithm::ExpandForward(GraphReader& graphreader,
     // Skip this edge if permanently labeled (best path already found to this
     // directed edge), if no access is allowed to this edge (based on costing method),
     // or if a complex restriction exists.
-    int has_time_restrictions = -1;
+    uint8_t has_time_restrictions = -1;
     if (current_es->set() == EdgeSet::kPermanent ||
         !current_costing->Allowed(directededge, pred, tile, edgeid, 0, 0, has_time_restrictions) ||
         current_costing->Restricted(directededge, pred, edgelabels_, tile, edgeid, true)) {
@@ -219,7 +219,7 @@ void AStarBSSAlgorithm::ExpandForward(GraphReader& graphreader,
     // Add to the adjacency list and edge labels.
     uint32_t idx = edgelabels_.size();
     edgelabels_.emplace_back(pred_idx, edgeid, directededge, newcost, sortcost, dist, mode, 0,
-                             transition_cost);
+                             transition_cost, baldr::kInvalidRestriction);
     *current_es = {EdgeSet::kTemporary, idx};
     adjacencylist_.add(idx);
   }
@@ -247,7 +247,7 @@ AStarBSSAlgorithm::GetBestPath(valhalla::Location& origin,
                                GraphReader& graphreader,
                                const sif::mode_costing_t& mode_costing,
                                const TravelMode mode,
-                               const Options& /*options*/) {
+                               const Options&) {
   // Set the mode and costing
   mode_ = mode;
   pedestrian_costing_ = mode_costing[static_cast<uint32_t>(TravelMode::kPedestrian)];
@@ -453,7 +453,7 @@ void AStarBSSAlgorithm::SetOrigin(GraphReader& graphreader,
     // of the path.
     uint32_t d = static_cast<uint32_t>(directededge->length() * (1.0f - edge.percent_along()));
     EdgeLabel edge_label(kInvalidLabel, edgeid, directededge, cost, sortcost, dist,
-                         TravelMode::kPedestrian, d, Cost{});
+                         TravelMode::kPedestrian, d, Cost{}, baldr::kInvalidRestriction);
     // Set the origin flag
     edge_label.set_origin();