+++ test/actor.cc
@@ -15,46 +15,10 @@ using namespace valhalla;
 
 namespace {
 
-boost::property_tree::ptree make_conf() {
-  // fake up config against pine grove traffic extract
-  auto conf = test::json_to_pt(R"({
-      "mjolnir":{"tile_dir":"test/traffic_matcher_tiles"},
-      "loki":{
-        "actions":["locate","route","sources_to_targets","optimized_route","isochrone","trace_route","trace_attributes","transit_available"],
-        "logging":{"long_request": 100},
-        "service_defaults":{"minimum_reachability": 50,"radius": 0,"search_cutoff": 35000, "node_snap_tolerance": 5, "street_side_tolerance": 5, "street_side_max_distance": 1000, "heading_tolerance": 60}
-      },
-      "thor":{"logging":{"long_request": 110}},
-      "skadi":{"actons":["height"],"logging":{"long_request": 5}},
-      "meili":{"customizable": ["breakage_distance"],
-               "mode":"auto","grid":{"cache_size":100240,"size":500},
-               "default":{"beta":3,"breakage_distance":2000,"geometry":false,"gps_accuracy":5.0,"interpolation_distance":10,
-               "max_route_distance_factor":3,"max_route_time_factor":3,"max_search_radius":100,"route":true,
-               "search_radius":50,"sigma_z":4.07,"turn_penalty_factor":200}},
-      "service_limits": {
-        "auto": {"max_distance": 5000000.0, "max_locations": 20,"max_matrix_distance": 400000.0,"max_matrix_locations": 50},
-        "auto_shorter": {"max_distance": 5000000.0,"max_locations": 20,"max_matrix_distance": 400000.0,"max_matrix_locations": 50},
-        "bicycle": {"max_distance": 500000.0,"max_locations": 50,"max_matrix_distance": 200000.0,"max_matrix_locations": 50},
-        "bus": {"max_distance": 5000000.0,"max_locations": 50,"max_matrix_distance": 400000.0,"max_matrix_locations": 50},
-        "hov": {"max_distance": 5000000.0,"max_locations": 20,"max_matrix_distance": 400000.0,"max_matrix_locations": 50},
-        "taxi": {"max_distance": 5000000.0,"max_locations": 20,"max_matrix_distance": 400000.0,"max_matrix_locations": 50},
-        "isochrone": {"max_contours": 4,"max_distance": 25000.0,"max_locations": 1,"max_time_contour": 120, "max_distance_contour":200},
-        "max_avoid_locations": 50,"max_radius": 200,"max_reachability": 100,"max_alternates":2,
-        "multimodal": {"max_distance": 500000.0,"max_locations": 50,"max_matrix_distance": 0.0,"max_matrix_locations": 0},
-        "pedestrian": {"max_distance": 250000.0,"max_locations": 50,"max_matrix_distance": 200000.0,"max_matrix_locations": 50,"max_transit_walking_distance": 10000,"min_transit_walking_distance": 1},
-        "skadi": {"max_shape": 750000,"min_resample": 10.0},
-        "trace": { "max_best_paths": 4, "max_best_paths_shape": 100, "max_distance": 200000.0, "max_gps_accuracy": 100.0, "max_search_radius": 100, "max_shape": 16000 },
-        "transit": {"max_distance": 500000.0,"max_locations": 50,"max_matrix_distance": 200000.0,"max_matrix_locations": 50},
-        "truck": {"max_distance": 5000000.0,"max_locations": 20,"max_matrix_distance": 400000.0,"max_matrix_locations": 50}
-      }
-    })");
-
-  conf.get_child("mjolnir").put("tile_dir", VALHALLA_SOURCE_DIR "test/traffic_matcher_tiles");
-  return conf;
-}
+// fake up config against pine grove traffic extract
+const auto conf = test::make_config(VALHALLA_SOURCE_DIR "test/traffic_matcher_tiles");
 
 TEST(Actor, Basic) {
-  auto conf = make_conf();
   tyr::actor_t actor(conf);
 
   actor.route(R"({"locations":[{"lat":40.546115,"lon":-76.385076,"type":"break"},
@@ -88,18 +52,9 @@ TEST(Actor, Basic) {
   // TODO: test the rest of them
 }
 
-class ActorInterrupt : public ::testing::Test {
-protected:
-  void SetUp() override {
-    conf = make_conf();
-  }
+struct test_exception_t {};
 
-  struct test_exception_t {};
-
-  boost::property_tree::ptree conf;
-};
-
-TEST_F(ActorInterrupt, Route) {
+TEST(Actor, Route) {
   tyr::actor_t actor(conf);
   std::string request = R"({"locations":[{"lat":40.546115,"lon":-76.385076,"type":"break"},
         {"lat":40.544232,"lon":-76.385752,"type":"break"}],"costing":"auto"})";
@@ -107,7 +62,7 @@ TEST_F(ActorInterrupt, Route) {
   EXPECT_THROW(actor.route(request, &interrupt), test_exception_t);
 }
 
-TEST_F(ActorInterrupt, TraceAttributes) {
+TEST(Actor, TraceAttributes) {
   tyr::actor_t actor(conf);
   std::string request = R"({"shape":[{"lat":40.546115,"lon":-76.385076},
         {"lat":40.544232,"lon":-76.385752}],"costing":"auto","shape_match":"map_snap"})";